// src/schema/user.schema.ts
import { z } from "zod";

export const userSchemaZod = z.object({
  name: z.string().min(2, "name must be at least 2 characters"),
  email: z.string().email("Invalid email"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});

export type userInput = z.infer<typeof userSchemaZod>;